fix: initialization process

+++ src/dao/MessageDao.ts
@@ -1,12 +1,12 @@
 import { Firestore } from "@google-cloud/firestore";
 import isNil from "lodash.isnil";
 import omit from "lodash.omitby";
-import { Chat, ChatDTO } from "../entities/Chat";
+import { Message, MessageDTO } from "../entities/Message";
 import { DaoError } from "../errors/dao";
-import { IChatDao } from "./IChatDao";
+import { IMessageDao } from "./IMessageDao";
 import { DaoTable } from "./IDao";
 
-export class ChatDao implements IChatDao {
+export class MessageDao implements IMessageDao {
   transaction!: FirebaseFirestore.Transaction;
 
   constructor(
@@ -14,124 +14,118 @@ export class ChatDao implements IChatDao {
     private readonly tableName: DaoTable
   ) {}
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  removeMember(_memberId: string): Promise<boolean> {
-    throw new Error("Method not implemented.");
-  }
-
-  async get(id: string, filter?: { adminId: string }): Promise<Chat | null> {
+  async get(id: string): Promise<Message | null> {
     try {
       const docRef = this.db.collection(this.tableName).doc(id);
-      let chatSnap: FirebaseFirestore.DocumentSnapshot<
+      let messageSnap: FirebaseFirestore.DocumentSnapshot<
         FirebaseFirestore.DocumentData,
         FirebaseFirestore.DocumentData
       >;
 
       if (this.transaction) {
-        chatSnap = await this.transaction.get(docRef);
+        messageSnap = await this.transaction.get(docRef);
       } else {
-        chatSnap = await docRef.get();
+        messageSnap = await docRef.get();
       }
 
-      const chat = chatSnap.data() as Chat | null;
+      const chat = messageSnap.data() as Message | null;
       if (!chat) {
         return null;
       }
 
-      if (filter?.adminId) {
-        if (chat.adminId !== filter.adminId) {
-          return null;
-        }
-      }
-
       return {
         ...chat,
-        id,
+        chatId: id,
       };
     } catch (error) {
       throw new DaoError({
-        name: "ChatDao",
+        name: "MessageDao",
         message: "Unable to retrieve chat",
-        id,
-        filter,
+        chatId: id,
         error,
       });
     }
   }
 
-  // TODO: apply pagination
-  async getAllByAdmin(adminId: string): Promise<Chat[]> {
+  async getAllChatMessages(
+    chatId: string,
+    filter?: { from?: string; to?: string }
+  ): Promise<Message[]> {
     try {
-      const docRef = this.db
+      let docRef = this.db
         .collection(this.tableName)
-        .where("adminId", "==", adminId)
+        .where("chatId", "==", chatId)
         .orderBy("createdAt", "desc");
 
-      let chatSnap: FirebaseFirestore.QuerySnapshot<
+      let messageSnap: FirebaseFirestore.QuerySnapshot<
         FirebaseFirestore.DocumentData,
         FirebaseFirestore.DocumentData
       >;
 
-      if (this.transaction) {
-        chatSnap = await this.transaction.get(docRef);
-      } else {
-        chatSnap = await docRef.get();
+      if (filter?.from) {
+        docRef = docRef.where("createdAt", ">=", filter.from);
       }
 
-      const chats = chatSnap.docs;
+      if (filter?.to) {
+        docRef = docRef.where("createdAt", "<=", filter.to);
+      }
+
+      if (this.transaction) {
+        messageSnap = await this.transaction.get(docRef);
+      } else {
+        messageSnap = await docRef.get();
+      }
+
+      const chats = messageSnap.docs;
       if (!chats.length) {
         return [];
       }
 
-      const allChats = chats.map((doc) => {
-        const chat = doc.data() as Chat;
+      const allMessages = chats.map((doc) => {
+        const chat = doc.data() as Message;
         return {
           ...chat,
           id: doc.id,
         };
       });
 
-      return allChats;
+      return allMessages;
     } catch (error) {
       throw new DaoError({
-        name: "ChatDao",
-        message: "Unable to retrieve chats",
-        adminId,
+        name: "MessageDao",
+        message: "Unable to retrieve message",
+        chatId,
         error,
       });
     }
   }
 
-  async save(chatDto: Partial<ChatDTO>, id?: string): Promise<Partial<Chat>> {
+  async create(chatDto: MessageDTO): Promise<Message> {
     try {
-      const payload = omit<ChatDTO>(chatDto, isNil);
-      const collectionRef = this.db.collection(this.tableName);
+      const payload = omit<MessageDTO>(chatDto, isNil);
+
       let chatRef: FirebaseFirestore.DocumentReference<
         FirebaseFirestore.DocumentData,
         FirebaseFirestore.DocumentData
       >;
 
-      if (id) {
-        chatRef = collectionRef.doc(id);
-      } else {
-        chatRef = collectionRef.doc();
-      }
-
       if (this.transaction) {
+        chatRef = this.db.collection(this.tableName).doc();
         this.transaction.set(chatRef, payload, { merge: true });
       } else {
-        await chatRef.set(payload, { merge: true });
+        await this.db.collection(this.tableName).add(payload);
       }
 
       return {
         ...chatDto,
-        isGroup: chatDto.isGroup || false,
-        members: chatDto.members || [],
+        mentionedIds: chatDto.mentionedIds || [],
+        command: chatDto.command || "",
+        sentTo: chatDto.sentTo || null,
       };
     } catch (error) {
       throw new DaoError({
-        name: "ChatDao",
-        message: "Unable to create chat",
+        name: "MessageDao",
+        message: "Unable to create message",
         chat: chatDto,
         error,
       });
@@ -140,11 +134,15 @@ export class ChatDao implements IChatDao {
 
   async delete(id: string): Promise<void> {
     try {
-      const data = (await this.db.collection(this.tableName).where("id", "==", id).get()).docs[0];
-      await this.db.collection(this.tableName).doc(data.id).delete();
+      const data = (await this.db.collection(this.tableName).where("chatId", "==", id).get())
+        .docs[0];
+
+      if (data) {
+        await this.db.collection(this.tableName).doc(data.id).delete();
+      }
     } catch (error) {
       throw new DaoError({
-        name: "ChatDao",
+        name: "MessageDao",
         message: "Unable to delete chat",
         id,
         error,
@@ -162,7 +160,7 @@ export class ChatDao implements IChatDao {
       }
     } catch (error) {
       throw new DaoError({
-        name: "ChatDao",
+        name: "MessageDao",
         message: "Unable to delete chat",
         id,
         error,
